[SIL] Explicitly use `uint64_t` for `maxBitfieldID`
The `maxBitfieldID` was defined as `size_t` and used to assert
bitfield ID values. The `currentBitfieldID` of `SILFunction` is
defined as `uint64_t`, so we should consistently use `uint64_t`
for `maxBitfieldID` as well for 32-bit platforms.

+++ include/swift/SIL/SILValue.h
@@ -1033,7 +1033,7 @@ class Operand {
 public:
   enum { numCustomBits = 8 };
 
-  constexpr static const size_t maxBitfieldID =
+  constexpr static const uint64_t maxBitfieldID =
       std::numeric_limits<uint64_t>::max() >> numCustomBits;
 
 private:
